+++ cfe/method/function/cf_cluster_mst.py
@@ -1,3 +1,5 @@
+from typing import Literal, Optional
+
 import anndata as ad
 import networkx as nx
 import numpy as np
@@ -6,26 +8,29 @@ import scanpy as sc
 from sklearn.metrics.pairwise import pairwise_distances
 
 
-def cf_cluster_mst(adata: ad.AnnData, prior_information: dict = {}, parameters: dict = {}):
-    # 1. prepare data
-    adata = adata.copy()
+def cluster_mst(
+    adata: ad.AnnData,
+    repreprocess: bool = True,
+    pca_ndim: int = 5,
+    basis: str = "X_pca",
+    recluster: bool = True,
+    cluster_key: str = "clusters",
+    distance_metric: Optional[Literal["euclidean", "cosine", "manhattan", "cityblock", "l1", "l2"]] = "euclidean",
+    **kwargs,
+):
+    # 1. preprocess
     adata.obs.reset_index(drop=True, inplace=True)
+    if repreprocess and (basis == "X_pca"):
+        sc.pp.pca(adata, n_comps=pca_ndim)
+    X_emb = adata.obsm[basis]
 
-    # 2. preprocess
-    sc.pp.pca(adata, n_comps=parameters["ndim"])
-    X_emb = adata.obsm["X_pca"]
-
-    # 3. execute method
+    # 2. execute method
     # (1) Cluster cells, with the center point as a milestone
-    if "groups_id" not in prior_information:
+    if recluster:
         # new cluster
         sc.pp.neighbors(adata)
         sc.tl.leiden(adata)
         cluster_key = "leiden"
-    else:
-        # cluster in prior_information
-        cluster_key = "mst_cluster"
-        adata.obs[cluster_key] = prior_information["groups_id"]
     adata.obs[cluster_key] = pd.Categorical(adata.obs[cluster_key])
     # (2) Calculate the low dimensional coordinates of the clustering centers
     centers = np.array(list(adata.obs.groupby(cluster_key).apply(lambda x: X_emb[list(x.index)].mean(axis=0))))
@@ -33,22 +38,36 @@ def cf_cluster_mst(adata: ad.AnnData, prior_information: dict = {}, parameters: 
     cluster_milestones = [milestone_ids[i] for i in adata.obs[cluster_key].cat.codes]
     centers = pd.DataFrame(centers, index=milestone_ids)
     # (3) Calculate the distance between cluster centers
-    distance_metric = parameters["distance_metric"]
     dis = pd.DataFrame(pairwise_distances(centers, metric=distance_metric), index=milestone_ids, columns=milestone_ids)
     disdf = pd.DataFrame(data=dis.unstack().reset_index().values, columns=["from", "to", "weight"])  # 转化为长数据
     # (4) Calculate the distance between milestones and construct the minimum spanning tree as the milestone network
     G = nx.from_pandas_edgelist(disdf, source="from", target="to", edge_attr="weight")
     mst = nx.minimum_spanning_tree(G, weight="weight")
 
-    # 4. extract results
+    # 3. extract results
     milestone_network = nx.to_pandas_edgelist(mst)
     milestone_network.rename(columns={"source": "from", "target": "to", "weight": "length"}, inplace=True)
     milestone_network["directed"] = False
 
-    # 5. save results
+    # 4. save results
     trajectory_dict = {
         "milestone_network": milestone_network,
         "cluster": cluster_milestones,
     }
 
     return trajectory_dict
+
+
+def cf_cluster_mst(
+    adata: ad.AnnData,
+    prior_information: dict = None,
+    parameters: dict = None,
+    **kwargs,
+):
+    if (prior_information is None) and (parameters is None):
+        # for new backend call, function(**kwargs)
+        return cluster_mst(adata, **kwargs)
+    else:
+        # for old backend call, function(prior_information, parameters)
+        parameters.update(prior_information)
+        return cluster_mst(adata, **parameters)