+++ cfe/method/function/cf_projection_mst.py
@@ -1,5 +1,5 @@
 #!/usr/local/bin/python3
-import pickle
+from typing import Literal, Optional
 
 import anndata as ad
 import networkx as nx
@@ -9,35 +9,35 @@ import scanpy as sc
 from sklearn.metrics.pairwise import pairwise_distances
 
 
-def cf_projection_mst(adata: ad.AnnData, prior_information: dict = {}, parameters: dict = {}):
-    # 1. prepare data
-    adata = adata.copy()
+def projection_mst(
+    adata: ad.AnnData,
+    repreprocess: bool = True,
+    pca_ndim: int = 5,
+    basis: str = "X_pca",
+    recluster: bool = True,
+    cluster_key: str = "clusters",
+    distance_metric: Optional[Literal["euclidean", "cosine", "manhattan", "cityblock", "l1", "l2"]] = "euclidean",
+):
+    # 1. preprocess
     adata.obs.reset_index(drop=True, inplace=True)
+    if repreprocess and (basis == "X_pca"):
+        sc.pp.pca(adata, n_comps=pca_ndim)
+    X_emb = adata.obsm[basis]
 
-    # 2. preprocess
-    sc.pp.pca(adata, n_comps=parameters["ndim"])
-    X_emb = adata.obsm["X_pca"]
-
-    # 3. execute method
-    # (1) Cluster cells, with the center point as a milestone
-    if "groups_id" not in prior_information:
+    # 2. execute method
+    # (1) if recluster cells, with the center point as a milestone
+    if recluster:
         # new cluster
         sc.pp.neighbors(adata)
         sc.tl.leiden(adata)
         cluster_key = "leiden"
-    else:
-        # cluster in prior_information
-        cluster_key = "mst_cluster"
-        adata.obs[cluster_key] = prior_information["groups_id"]
     # (2) Calculate the low dimensional coordinates of the clustering centers
     centers = np.array(list(adata.obs.groupby(cluster_key).apply(lambda x: X_emb[list(x.index)].mean(axis=0))))
     milestone_ids = [f"M{i}" for i in range(centers.shape[0])]
     centers = pd.DataFrame(centers, index=milestone_ids)
     # (3) Calculate the distance between cluster centers
-    distance_metric = parameters["distance_metric"]
     dis = pd.DataFrame(pairwise_distances(centers, metric=distance_metric), index=milestone_ids, columns=milestone_ids)
-    disdf = pd.DataFrame(data=dis.unstack().reset_index().values, columns=["from", "to", "weight"])  # 转化为长数据
-
+    disdf = pd.DataFrame(data=dis.unstack().reset_index().values, columns=["from", "to", "weight"])  # width data to long data
     # (4) Calculate the distance between milestones and construct the minimum spanning tree as the milestone network
     G = nx.from_pandas_edgelist(disdf, source="from", target="to", edge_attr="weight")
     mst = nx.minimum_spanning_tree(G, weight="weight")
@@ -45,13 +45,13 @@ def cf_projection_mst(adata: ad.AnnData, prior_information: dict = {}, parameter
     milestone_network.rename(columns={"source": "from", "target": "to", "weight": "length"}, inplace=True)
     milestone_network["directed"] = False
 
-    # 4. extract results
+    # 3. extract results
     comp_ids = [f"comp_{i+1}" for i in range(centers.shape[1])]
     X_emb = pd.DataFrame(X_emb, index=adata.obs.index, columns=comp_ids)
     milestone_emb = centers
     milestone_emb.columns = comp_ids
 
-    # 5. save results
+    # 4. save results
     trajectory_dict = {
         "milestone_network": milestone_network,
         "X_emb": X_emb,
@@ -60,13 +60,16 @@ def cf_projection_mst(adata: ad.AnnData, prior_information: dict = {}, parameter
     return trajectory_dict
 
 
-if __name__ == "__main__":
-    from parse_args import parse_args
-
-    adata, prior_information, parameters, output_filename = parse_args()
-
-    trajectory_dict = cf_projection_mst(adata, prior_information, parameters)
-
-    with open(output_filename, "wb") as f:
-        pickle.dump(trajectory_dict, f)
-    print("projection_mst Finish!")
+def cf_projection_mst(
+    adata: ad.AnnData,
+    prior_information: dict = None,
+    parameters: dict = None,
+    **kwargs,
+):
+    if (prior_information is None) and (parameters is None):
+        # for new backend call, function(**kwargs)
+        return projection_mst(adata, **kwargs)
+    else:
+        # for old backend call, function(prior_information, parameters)
+        parameters.update(prior_information)
+        return projection_mst(adata, **parameters)