+++ cfe/method/function/cf_velovi.py
@@ -2,34 +2,38 @@ import anndata as ad
 import scvelo as scv
 
 
-def cf_velovi(adata: ad.AnnData, prior_information: dict = {}, parameters: dict = {}):
+def velovi(
+    adata: ad.AnnData,
+    repreprocess: bool = True,
+    filter_and_normalize_kwargs: dict = {},
+    moments_kwargs: dict = {},
+    velovi_model_kwargs: dict = {},
+    velovi_train_kwargs: dict = {},
+    n_sample: int = 25,
+):
+    # ref: https://docs.scvi-tools.org/en/stable/tutorials/notebooks/scrna/velovi.html
+    # the package is not available in cfe envirionment, so we import it here
     import scvi
 
-    # import torch
-    # cluster_key = prior_information.get("cluster_key", "clusters")  # do nothing, only prior information demo
-    max_epochs = parameters.get("max_epochs", 50)
+    # 1. preprocess
+    if repreprocess:
+        scv.pp.filter_and_normalize(adata, **filter_and_normalize_kwargs)
+        scv.pp.moments(adata, **moments_kwargs)
 
-    # 1. prepare data
-    adata = adata.copy()
-
-    # 2. preprocess
-    scv.pp.filter_and_normalize(adata, min_shared_counts=20, n_top_genes=2000)
-    scv.pp.moments(adata, n_pcs=30, n_neighbors=30)
-
-    # 3. execute method
+    # 2. execute method
     VELOVI = scvi.external.VELOVI  # extract the VELOVI class
     VELOVI.setup_anndata(adata, spliced_layer="Ms", unspliced_layer="Mu")
-    vae = VELOVI(adata)
-    vae.train(max_epochs=max_epochs)  # TODO: very slow, need GPU
+    vae = VELOVI(adata, **velovi_model_kwargs)
+    vae.train(**velovi_train_kwargs)  # TODO: very slow, need GPU
     # extract velocity to adata.layers["velocity"]
-    latent_time = vae.get_latent_time(n_samples=25)
-    velocities = vae.get_velocity(n_samples=25, velo_statistic="mean")
+    latent_time = vae.get_latent_time(n_samples=n_sample)
+    velocities = vae.get_velocity(n_samples=n_sample, velo_statistic="mean")
     t = latent_time
     scaling = 20 / t.max(0)
     adata.layers["velocity"] = velocities / scaling
     scv.tl.velocity_graph(adata)
 
-    # 4. extract results
+    # 3,4. extract and save results
     trajectory_dict = {
         "velocity": adata.layers["velocity"],
         "velocity_graph": adata.uns["velocity_graph"],
@@ -40,3 +44,18 @@ def cf_velovi(adata: ad.AnnData, prior_information: dict = {}, parameters: dict 
     }
 
     return trajectory_dict
+
+
+def cf_velovi(
+    adata: ad.AnnData,
+    prior_information: dict = None,
+    parameters: dict = None,
+    **kwargs,
+):
+    if (prior_information is None) and (parameters is None):
+        # for new backend call, function(**kwargs)
+        return velovi(adata, **kwargs)
+    else:
+        # for old backend call, function(prior_information, parameters)
+        parameters.update(prior_information)
+        return velovi(adata, **parameters)