+++ cfe/plot/__init__.py
@@ -1,3 +1,4 @@
+from .plot_benchmark import style_benchmark
 from .plot_embedding import plot_embedding_plotly
 from .plot_graph import plot_graph
 from .plot_topology import plot_topology
@@ -5,11 +6,4 @@ from .plot_trajectory import plot_trajectory
 from .plot_trajectory_new import plot_trajectory_new
 from .plot_wrapper import plot_wrapper
 
-__all__ = [
-    "plot_trajectory",
-    "plot_graph",
-    "plot_topology",
-    "plot_wrapper",
-    "plot_embedding_plotly",
-    "plot_trajectory_new",
-]
+__all__ = ["plot_trajectory", "plot_graph", "plot_topology", "plot_wrapper", "plot_embedding_plotly", "plot_trajectory_new", "style_benchmark"]