+++ cfe/util/__init__.py
@@ -1,5 +1,12 @@
 from .available_metadata import get_available_method_df
+from .parse_benchmark import parse_resource_useage_string
 from .project import project_to_segments
 from .random_time_string import parse_random_time_string, random_time_string
 
-__all__ = ["get_available_method_df", "random_time_string", "parse_random_time_string", "project_to_segments"]
+__all__ = [
+    "get_available_method_df",
+    "random_time_string",
+    "parse_random_time_string",
+    "project_to_segments",
+    "parse_resource_useage_string",
+]