+++ cfe/util/docs.py
@@ -1,0 +1,3 @@
+def merge_doc(doc1, doc2, head_keep=2, return_keep=1):
+    # TODO: merge parameters of doc1 and doc2, keep desciption of doc2 and return result of doc1 as combine result
+    pass