+++ tests/method/function/test_cf_paga.py
@@ -12,11 +12,28 @@ class TestCFPAGA:
         self.fadata = cfe.data.FateAnnData.from_anndata(adata)
         self.fadata.obs.index = self.fadata.obs["cell_id"].tolist()
 
-    def test_paga(self):
+    def test_paga_old(self):
         # add priority and parameeters
-        prior_information = {"start_id": "cell1", "groups_id": self.fadata.obs["lineage"].tolist()}
-        parameters = {"connectivity_cutoff": 0.8, "filter_features": False}
+        prior_information = {"start_id": "cell1"}
+        parameters = {
+            "cluster_key": "lineage",
+            "repreprocess": True,
+            "filter_and_normalize_kwargs": {},
+            "neighbors_kwargs": {},
+            "n_dcs": 2,
+            "connectivity_cutoff": 0.5,
+        }
         trajectory_dict = cfe.method.cf_paga(self.fadata, prior_information, parameters)
+        assert trajectory_dict.keys() == {"branch_network", "branches", "branch_progressions"}
+
+    def test_paga_new(self):
+        # add priority and parameeters
+        parameters = {
+            "start_id": "cell1",
+            "cluster_key": "lineage",
+            "connectivity_cutoff": 0.5,
+        }
+        trajectory_dict = cfe.method.cf_paga(self.fadata, **parameters)
         assert trajectory_dict.keys() == {"branch_network", "branches", "branch_progressions"}