+++ tests/method/function/test_cf_scvelo.py
@@ -12,13 +12,28 @@ class TestCFscVelo:
         self.fadata = cfe.data.FateAnnData.from_anndata(adata)
         self.fadata.obs.index = self.fadata.obs["cell_id"].tolist()
 
-    def test_scvelo(self):
+    def test_scvelo_old(self):
         # add priority and parameeters
-        prior_information = {
-            "cluster_key": "lineage",
+        prior_information = {}
+        parameters = {
+            "repreprocess": True,
+            "filter_and_normalize_kwargs": {
+                "min_shared_counts": 20,
+                "n_top_genes": 2000,
+            },
+            "moments_kwargs": {
+                "n_pcs": 20,
+                "n_neighbors": 10,
+            },
+            "velocity_kwargs": {},
+            "velocity_graph_kwargs": {},
         }
-        parameters = {}
         trajectory_dict = cfe.method.cf_scvelo(self.fadata, prior_information, parameters)
+        assert trajectory_dict.keys() == {"velocity", "velocity_graph", "velocity_graph_neg", "neighbors", "obs_index", "var_index"}
+
+    def test_scvelo_new(self):
+        parameters = {}
+        trajectory_dict = cfe.method.cf_scvelo(self.fadata, **parameters)
         assert trajectory_dict.keys() == {"velocity", "velocity_graph", "velocity_graph_neg", "neighbors", "obs_index", "var_index"}