+++ tests/method/test_fate_conda_backend.py
@@ -39,6 +39,18 @@ class TestCondaBackend:
         self.conda_backend.run(fadata, parameters)
         assert fadata.is_wrapped_with_trajectory
 
+    def test_call(self):
+        adata = sc.read(f"{os.path.dirname(__file__)}/../data/bifurcating.h5ad")
+        adata.obs.index = adata.obs["cell_id"]
+
+        parameters = {
+            "start_id": "cell1",
+            "cluster_key": "lineage",
+            "connectivity_cutoff": 0.5,
+        }
+
+        self.function_backend(adata, **parameters)
+
     def test_load_definition(self):
         # _load_definition is called in test_load_backend, which is called in __init__
         definition = self.conda_backend.definition