+++ docs/change_log.md
@@ -4,4 +4,8 @@
 
 ## [2024.12.24: 0.1.0 version, project framework finish](change_log/2024.12.24.md)
 
-## [2024.12.31 0.1.1 version, method selection](change_log/2025.01.19.md)
+## [2024.12.31: 0.1.1 version, method selection](change_log/2025.01.19.md)
+
+## [2025.07.02: 0.1.2 Framework core: method, metric, plot](change_log/2025.07.02.md)
+
+## [dev](change_log/dev.md)